Add minibasket to navbar
Adds a "minibasket" dropdown to the navbar with item count and basket
price shown on the button. Also assigns an alias to the shopBasket
component used on the payment page to prevent clashes, adds Bootstrap's
fonts that I missed earlier for the button's icon, and finally specifies
a couple more parameters added for the two-components-on-one-page thing.